Use size_t instead of unsigned int

+++ arrows/ceres/types.cxx
@@ -45,7 +45,7 @@ namespace kwiver {                                                     \
   typedef NS::ceres_type T;                                            \
   std::string options_str = "\nMust be one of the following options:"; \
   std::string opt;                                                     \
-  for( unsigned i = 0; i < 20; ++i )                                   \
+  for( size_t i = 0; i < 20; ++i )                                     \
   {                                                                    \
     opt = NS::ceres_type##ToString( static_cast< T >( i ) );           \
     if( opt == "UNKNOWN" )                                             \
