+++ arrows/tests/test_image.h
@@ -36,11 +36,11 @@ populate_vital_image( kwiver::vital::image& img, T minv, T maxv )
   const double range = static_cast< double >( maxv ) -
                        static_cast< double >( minv );
   const double offset = -static_cast< double >( minv );
-  for( unsigned int p = 0; p < img.depth(); ++p )
+  for( size_t p = 0; p < img.depth(); ++p )
   {
-    for( unsigned int j = 0; j < img.height(); ++j )
+    for( size_t j = 0; j < img.height(); ++j )
     {
-      for( unsigned int i = 0; i < img.width(); ++i )
+      for( size_t i = 0; i < img.width(); ++i )
       {
         img.at< T >(
           i, j,
@@ -66,12 +66,12 @@ populate_vital_image( kwiver::vital::image& img )
 }
 
 // Parameters for common test of get_image function
-constexpr unsigned int full_width = 60;
-constexpr unsigned int full_height = 40;
-constexpr unsigned int cropped_width = 30;
-constexpr unsigned int cropped_height = 20;
-constexpr unsigned int x_offset = 10;
-constexpr unsigned int y_offset = 5;
+constexpr size_t full_width = 60;
+constexpr size_t full_height = 40;
+constexpr size_t cropped_width = 30;
+constexpr size_t cropped_height = 20;
+constexpr size_t x_offset = 10;
+constexpr size_t y_offset = 5;
 
 // ----------------------------------------------------------------------------
 // helper function to generate common test of get_image funcion
