+++ vital/algo/compute_depth.h
@@ -51,7 +51,7 @@ public:
     std::vector< kwiver::vital::image_container_sptr > const& frames,
     std::vector< kwiver::vital::camera_perspective_sptr > const& cameras,
     double depth_min, double depth_max,
-    unsigned int reference_frame,
+    size_t reference_frame,
     vital::bounding_box< int > const& roi,
     std::vector< kwiver::vital::image_container_sptr > const& masks =
     std::vector< kwiver::vital::image_container_sptr >( ) ) const;
@@ -77,7 +77,7 @@ public:
     std::vector< kwiver::vital::image_container_sptr > const& frames,
     std::vector< kwiver::vital::camera_perspective_sptr > const& cameras,
     double depth_min, double depth_max,
-    unsigned int reference_frame,
+    size_t reference_frame,
     vital::bounding_box< int > const& roi,
     kwiver::vital::image_container_sptr& depth_uncertainty,
     std::vector< kwiver::vital::image_container_sptr > const& masks =
@@ -86,7 +86,7 @@ public:
   /// Typedef for the callback function signature
   typedef std::function< bool ( kwiver::vital::image_container_sptr,
                                 std::string const&,
-                                unsigned int,
+                                size_t,
                                 kwiver::vital::image_container_sptr ) >
     callback_t;
